training/requirements.txt
stable-baselines3[extra]>=2.3.0
gymnasium>=0.29.0
pyzmq>=25.0.0
numpy>=1.24.0
torch>=2.0.0
tensorboard>=2.14.0
modal>=0.64.0
